+++ CHANGELOG.md
@@ -10,6 +10,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- `BlockstackSession.validateProofs` method and `Proof` object to verify social acounts.
+- `BlockstackSession.getAppBucketUrl` method to retrieve the user's bucket url of the app.
+- `BlockstackSession.getUserAppFileUrl` method to retrieve a user's file of an app with a given path.
+- `UserData.hubUrl` property for use with `getUserAppFileUrl`
+
+### Changed
+- `BlockstackSession.makeAuthResponse` has a new parameter `privateAppKey`
+- improved integration tests
+
 ## [0.3.0] - 2018-07-27
 
 ### Added